requirements.txt
PyPDF2~=2.2.0
python-docx~=0.8.11
pywin32~=304